src/app/requirements.txt
gradio>=4.0
databricks-sdk
requests
